+++ lib/nylas/account.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Nylas
   # Representation of the accounts for Account management purposes.
   # @see https://docs.nylas.com/reference#account-management
@@ -21,6 +23,13 @@ module Nylas
 
     def downgrade
       response = execute(method: :post, path: "#{resource_path}/downgrade")
+      response[:success]
+    end
+
+    def revoke_all(keep_access_token: nil)
+      payload = JSON.dump(keep_access_token: keep_access_token) if keep_access_token
+
+      response = execute(method: :post, path: "#{resource_path}/revoke-all", payload: payload)
       response[:success]
     end