+++ lib/nylas/api.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Nylas
   # Methods to retrieve data from the Nylas API as Ruby objects
   class API
@@ -15,20 +17,23 @@ module Nylas
     #                            you're using a self-hosted Nylas instance.
     # @param service_domain [String] (Optional) Host you are authenticating OAuth against.
     # @return [Nylas::API]
-    # rubocop:disable Metrics/ParameterLists
     def initialize(client: nil, app_id: nil, app_secret: nil, access_token: nil,
                    api_server: "https://api.nylas.com", service_domain: "api.nylas.com")
       self.client = client || HttpClient.new(app_id: app_id, app_secret: app_secret,
                                              access_token: access_token, api_server: api_server,
                                              service_domain: service_domain)
     end
-    # rubocop:enable Metrics/ParameterLists
 
     # @return [String] A Nylas access token for that particular user.
-    def authenticate(name:, email_address:, provider:, settings:, reauth_account_id: nil)
-      NativeAuthentication.new(api: self).authenticate(name: name, email_address: email_address,
-                                                       provider: provider, settings: settings,
-                                                       reauth_account_id: reauth_account_id)
+    def authenticate(name:, email_address:, provider:, settings:, reauth_account_id: nil, scopes: nil)
+      NativeAuthentication.new(api: self).authenticate(
+        name: name,
+        email_address: email_address,
+        provider: provider,
+        settings: settings,
+        reauth_account_id: reauth_account_id,
+        scopes: scopes
+      )
     end
 
     # @return [Collection<Contact>] A queryable collection of Contacts
@@ -94,6 +99,14 @@ module Nylas
       response.code == 200 && response.empty?
     end
 
+    # Returns list of IP addresses
+    # @return [Hash]
+    # hash has keys of :updated_at (unix timestamp) and :ip_addresses (array of strings)
+    def ip_addresses
+      path = "/a/#{app_id}/ip_addresses"
+      client.as(client.app_secret).get(path: path)
+    end
+
     # @param message [Hash, String, #send!]
     # @return [Message] The resulting message
     def send!(message)
@@ -119,8 +132,11 @@ module Nylas
       @webhooks ||= Collection.new(model: Webhook, api: as(client.app_secret))
     end
 
-    private def prevent_calling_if_missing_access_token(method_name)
+    private
+
+    def prevent_calling_if_missing_access_token(method_name)
       return if client.access_token && !client.access_token.empty?
+
       raise NoAuthToken, method_name
     end
   end