+++ lib/nylas/contact.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Nylas
   # ActiveModel compliant interface for interacting with the Contacts API
   # @see https://docs.nylas.com/reference#contacts
@@ -29,6 +31,7 @@ module Nylas
     attribute :office_location, :string
     attribute :notes, :string
     attribute :web_page, :web_page
+    attribute :source, :string
 
     has_n_of_attribute :groups, :contact_group
     has_n_of_attribute :emails, :email_address
@@ -41,6 +44,7 @@ module Nylas
     # to retrieve it from nylas every time.
     def picture
       return @picture_tempfile if @picture_tempfile
+
       @picture_tempfile = Tempfile.new
       @picture_tempfile.write(api.get(path: "#{resource_path}/picture"))
       @picture_tempfile.close