+++ lib/nylas/event.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Nylas
   # Structure to represent a the Event Schema.
   # @see https://docs.nylas.com/reference#events
@@ -23,7 +25,7 @@ module Nylas
     attribute :read_only, :boolean
     attribute :status, :string
     attribute :title, :string
-    attribute :when, :timespan
+    attribute :when, :when
     attribute :original_start_time, :unix_timestamp
 
     def busy?