+++ lib/nylas/file.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Nylas
   # Structure to represent a the File Schema.
   # @see https://docs.nylas.com/reference#events
@@ -22,12 +24,14 @@ module Nylas
     # @return [Tempfile] - Local copy of the file
     def download
       return file if file
+
       self.file = retrieve_file
     end
 
     # Redownloads a file even if it's been cached. Closes and unlinks the tempfile to help memory usage.
     def download!
       return download if file.nil?
+
       file.close
       file.unlink
       self.file = nil
@@ -40,16 +44,19 @@ module Nylas
 
     def save
       raise ModelNotUpdatableError if persisted?
+
       response = api.execute(path: "/files", method: :post, headers: { multipart: true },
                              payload: { file: file })
       attributes.merge(response.first)
       true
     end
 
-    private def retrieve_file
+    private
+
+    def retrieve_file
       response = api.get(path: "#{resource_path}/download")
       filename = response.headers.fetch(:content_disposition, "").gsub("attachment; filename=", "")
-      temp_file = Tempfile.new(filename)
+      temp_file = Tempfile.new(filename, :encoding => 'ascii-8bit')
       temp_file.write(response.body)
       temp_file.seek(0)
       temp_file