+++ spec/nylas/api_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "spec_helper"
 
 # This spec is the only one that should have any webmock stuff going on, everything else should use the
@@ -20,7 +22,13 @@ describe Nylas::API do
                                                        "No access token was provided and the " \
                                                        "current_account method requires one"
     end
+
+    it "sets X-Nylas-Client-Id header" do
+      client = Nylas::HttpClient.new(app_id: "not-real", app_secret: "also-not-real")
+      expect(client.default_headers).to include("X-Nylas-Client-Id" => "not-real")
+    end
   end
+
   describe "#execute" do
     it "builds the URL based upon the api_server it was initialized with"
     it "adds the nylas headers to the request"