+++ spec/spec_helper.rb
@@ -1,9 +1,14 @@
+# frozen_string_literal: true
+
 require "simplecov"
 SimpleCov.start
 
 require "nylas"
 require "pry"
 require "webmock/rspec"
+require "rspec-json_matcher"
+
+RSpec.configuration.include RSpec::JsonMatcher
 
 class FakeAPI
   def execute(method:, path:, payload: nil)